Deprecate helper methods of "smart-pointer" classes
The helper methods, which are so simple they merely assign the pointer/
handle to the class instance, were simply marked depreciated.

The helper methods, which require extra variable to store the output
pointer/handle before assigning it to the class instance, were replaced
with function overloads named by the original function helpers initially
wrapped. This allows easy porting of the legacy code to use WinStd
classes.

This commit also splits the helper functions to A and W (_UNICODE)
variants.

Signed-off-by: Simon Rozman <simon@rozman.si>